Desktop Access: discovery improvements (#53067)
* Desktop Access: support multiple discovery configs per-agent

Teleport's LDAP-based desktop discovery allows users to specify
a single search path (base DN) along with some filters and attributes.
This is suboptimal in cases where you want to discover desktops from
different areas of the directory hierarchy, as the only option is
to move the base DN to the first common ancestor and hope to filter
down the set from there.

This commit allows the configuration of multiple independent discovery
policies by exposing a new `discovery_configs` setting in the
configuration. Old configs that only set `discovery` will continue to
work, but it is illegal to set both `discovery` and `discovery_configs`.

Closes #43354

* Make discovery interval configurable

Closes #54246

* Add support for custom ports with LDAP discovery

Closes #11734