[v17] Add .pnpm-store to .gitignore (#54739)
* Add .pnpm-store to .gitignore

This adds .pnpm-store to .gitignore. This is generated in the project
root during dockerized builds and can negatively impact editor and
shell completion due to the very large directory structure; adding it
to .gitignore mostly fixes both of these issues.

To reproduce (without this change):

```
make -C build.assets release-amd64
```

Observe .pnpm-store created in the project root, and that fish tab
completion is functionally unusable.

* clean-ui: Remove .pnpm-store

* Move removing .pnpm-store to `make -C build.assets clean` instead

---------

Co-authored-by: Rafał Cieślak <rafal.cieslak@goteleport.com>