MWI: Bound Keypair Joining - Minimal Client Implementation (#54372)
* MWI: Bound Keypair Joining - Minimal Client Implementation

This contains a minimal client-side implementation of bound keypair
joining for `tbot`. This first iteration supports only preregistered
keys, and so adds a new `tbot keypair create` subcommand that
generates a keypair and prints a public key that can be copied into
a token.

In an effort to still take advantange of tbot's secure filesystem
abstraction and to not permanently tie the new join method to it,
this includes a minimal filesystem wrapper to adapt tbot's
destinations to a generic file read/write interface.

* Rebase join client to handle interactive rotation messages

This updates the client implementation to handle the revised challenge
handler signature. Rotation still isn't implemented (see TODOs), but
the message handler now explicitly accounts for rotation-type requests
from the server.

* Fix breakage after upstream changes, idempotent key generation

The `keypair generate` command is now idempotent and will not generate
a new keypair if one already exists in the specified destination.

Also, added a `--format=json` flag to print the key as
machine-readable JSON.

* Remove unnecessary whitespace

* Small TODO and doc comment updates

* Fix lints

* Mark `keypair create` as a hidden command for now

This hides it from `tbot help` so we can opt to tweak the CLI contract
before fully releasing the feature. It will be unhidden in the near
future.

* Fix typo