Terraform: Fix Access List import (#54803)
The most important bit is set tfschema.TimeValue.Format in CopyToTimestamp so non-nil times are
parsed correctly. Otherwise the import results in `parsing time "" as "2006-01-02T15:04:05Z07:00":
cannot parse "" as "2006"`. That affected all resources, because all of them have
`metadata.expires` field, but AccessList was hit the most because it has
`spec.audit.next_audit_date` Timestamp field.

The other part is marking more header fields as computed so the `apply` after the `import`` goes
smoothly and is less confusing.