Fix malformed redirect
This change fixes a redirect with a malformed destination path and
source path. The source path needs to have a single instance of the
`join-services-to-your-cluster` segment, since this is a section index
page. The destination path erroneously uses the beginning of a local
file path, and needs to omit the base `/docs/` segment.