Aggregate health status in unified resource cache
The health status for a resource with multiple servers, as in a database
proxied by multiple agents, will now be aggregated in the unified
resource cache. If there are multiple statuses and the statuses are not
all the same, then the aggregate status will be set to "mixed".
It is then the client's responsibility to list the servers for a "mixed"
status resource if they want details about individual resource server
health.