Remove superflous DeleteAll implementations (#6530)
This is a precursor to https://github.com/gravitational/teleport/pull/54867.
All DeleteAllFoo methods that existed solely for the cache which
are not exposed via a public API are being removed. To allow that
all code in enterprise must be converted first to use an alternate
method of cleaning up.

Source-commit: gravitational/teleport.e@f1004ff54134711f1f7fbde5e42456ed373cd6a4
Source-pull-requests: gravitational/teleport.e#6530