Refactor prerequisite partials
We currently have a number of partials with names similar to
`[a-z]+-prereqs-tabs` that include standard prerequisites for admin
guides. This change simplifies these partials by condensing them into a
single partial, the original `edition-prereqs-tabs.mdx` partial, and
using parameters to allow for variations.

The standard prerequisites partial includes installation steps for
client tools (Closes #54888).

This change also clarifies compatibility requirements for client tools
in the standard prerequisites partial. (Closes #51942)

Finally, this change mentions Community Edition in the Get Started
guide. This saves us from having to branch between commercial and
license-agnostic requirements in `edition-prereqs-tabs.mdx`, which can
just link to `get-started.mdx`. It also makes the Linux demo guide more
visible for practitioners who want to start with a demo.

Note that this change does not attempt to factor out the `tsh`
installation instructions from the `install-tsh.mdx` partial and the
`edition-prereqs-tabs.mdx` partial. This is because the docs engine does
not evaluate the `(! !)` include syntax if that syntax is within a list
item in another partial. Once we move to the Docusaurus-native partial
system, we can consider further refactoring.

Other changes:
- Remove homebrew tab from `install-tsh.mdx`. We don't need a separate
  tab for the homebrew warning.
- Factor out `tctl-tsh-prerequisite.mdx`. For the most part, we can do
  this by using `edition-prereqs-tabs.mdx`. In the self-signed certs
  guide, we need to inline the instructions since the Teleport edition
  requirements differ from the ones in `edition-prereqs-tabs.mdx`.