Update IC tests to clone proto directly (#6542)
The tests were making use of CloneResource() to clone an AWS Identity
Center Account. Said method is intended for use in the Unified Resource
Cache and is being phased out. The tests now make use of proto.Clone
directly on the underling message instead allowing the method to
be removed in OSS.

Source-commit: gravitational/teleport.e@b4497d88e8066e6174e3c22fabcf75599e5c9f57
Source-pull-requests: gravitational/teleport.e#6542