Convert Identity Center Accounts to Apps in the UnifiedResourceCache (#54911)
The resources are only ever presented to the UI as Apps so there is
room to simplify the special handling of Identity Center Accounts
stored in the URC. The extra machinery needed to convert the
identitycenterv1.Account into a types.ResourceWithLabels is no longer
needed and is removed. The IdentityCenterAccountMatcher was extended
to allow being created from an AppServer so that RBAC is still
enforced appropriately on any AppServers that are an IC Account.

The IdentityCenterAccountAssignments were also removed entirely
from the unified resource cache because they were never consumed.