[iac] github plugin implementation (#6545)
* [iac] github plugin implementation

Introduce GitHub plugin and service for audit log and metadata ingestion

This commit adds the initial implementation of the GitHub plugin and its associated service. The service is responsible for integrating with GitHub's API to periodically retrieve and forward relevant data to the Access Graph.

Key functionalities include:

- Polling GitHub audit logs in a paginated manner and forwarding them to Access Graph for further processing and correlation.

- Fetching and forwarding GitHub tokens, organization members, and their associated roles to enrich identity and access context in the Access Graph.

This integration lays the groundwork for deeper visibility into GitHub activity and access patterns within the broader system.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* handle review feedback

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

Source-commit: gravitational/teleport.e@279d0ae071044c540c44d1eb740160e2abb8f28e
Source-pull-requests: gravitational/teleport.e#6545