Register unimplemented secreportsv1 RPC service in OSS builds (#55056)
Updates https://github.com/gravitational/teleport/issues/54520.

Ensures that a secreportsv1 RPC service is always registered to
allow caches to become healthy. Without this depending on the build
type interacting with the secreports RPC service resulted in errors
about an unregistered service. The cache will be updated after the
accompanying change is made in teleport.e to ensure the service
is always registered their regardless of whether the access monitoring
feature is enabled.