Remove / prefix from ListUserGroups next token
(ServerWithRoles) ListUserGroups was returning the next page token
with a / preceeding the name of the group. When switching over to
the new cache collection, the cache was not accounting for this
and would result in an infinite loop. The prefix was a result of
calling backend.NextPaginationKey instead of providing the unique
group name.

This was not a problem in the past because the local UserGroup
storage service was joining the next token with the key prefix which
correctly built the key /user_groups/<next>. The new cache collection
is now trimming the `/` prefix to prevent the infinite looping.

The AuthServer implementation was also updated to no longer include
the prefix, which should be backward compatibile because of the
way the local services are constructing the backend key. Additionally,
tests were added to both the new caching layer and local storage
layer to validate both work if the nextKey provided has a `/` prefix
or not.

backend.NextPaginationKey was also removed since it was no longer
being used.