Allow unknown ListUnifiedResourceRequest.Kinds (#55211)
Unknown kinds resulted in the request erroring out and returning
an error about unknown kinds. This causes compatibility issues when
new kinds are added, i.e. git_server in v17 with a v16 leaf cluster
will prevent the leaf cluster resources page from loading in the
root web ui.

The errors are now treated as if access for unknown kinds are
denied which will filter them out and allow any known resources
to still be processed and returned.