Add Rust version override to universal macOS build. (#6606)
* Add Rust version override to universal macOS build.

This is necessary to ensure that the correct version of Rust is used when building the universal binary for macOS. It seems that the default Rust version (currently 1.87) is not compatible with our current build process.

The builds for our other platforms already have this override, so this change brings the universal macOS workflow in line with those.

* Defaulting rust target arch for universal

* Using rustup-set-version instead

Source-commit: gravitational/teleport.e@155decddc1473488df1c51e0042f59bdb0da3ae5
Source-pull-requests: gravitational/teleport.e#6606