Allow tunnel nodes to enable the ssh listener (#54215)
This allows the service to be connectable by users with direct
network access. All connections still require a valid user
certificate to be presented and will not permit any extra access.
This is intended to provide an optional connection path to hosts
that may provide reduced latency if the Proxy is not co-located with
the user and service.