Consolidate CI jobs
Combines all workflow jobs which test/lint/validate the changes of
a PR into a single CI workflow comprised of multiple jobs. The
motivation is to trigger only the required jobs based on the
changes of a PR in an efficient manner. Prior to this each workflow
was independently checking for the same set of changes and acting
appropriately. This also removes the remaining bypass jobs since
everything is now. While this might not reduce the amount of time
needed to merge a PR it may produce some monetary savings by no
longer running N workflows that check for the same change set to
ultimately determine that all N workflows may be skipped.