[iac] prevent a deadlock by calling `stream.Send` concurrently (#6639)
This PR fixes a deadlock caused by parallel calls to gRPC's
bidirectional `stream.Send`.

gRPC underlying `SendMsg` is not thread-safe which causes at least one
of the Send calls to never receive the acknowledge from the server.

For access graph, we run bulk-exporter with concurrency = 3 and we also
have the root goroutine trying to send dates sync.

This PR refactors that code and transforms the stream.Send to run in a
single goroutine.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

Source-commit: gravitational/teleport.e@c90fb90a10a62e1057c7a2fbae8d53f248aa62af
Source-pull-requests: gravitational/teleport.e#6639