Aggregate health status in unified resource cache (#54868)
* Aggregate health status in unified resource cache

The health status for a resource with multiple servers, as in a database
proxied by multiple agents, will now be aggregated in the unified
resource cache. If there are multiple statuses and the statuses are not
all the same, then the aggregate status will be set to "mixed".
It is then the client's responsibility to list the servers for a "mixed"
status resource if they want details about individual resource server
health.

* Update BenchmarkListUnifiedResources

* Improve aggregate health status cache performance

* reduce support for health predicate expressions to only support
  expressions on the "health.status" field. This reduces copying of the
  entire TargetHealth struct and eliminates the dynamic json tag lookup
  for health status matching.
* add GetTargetHealthStatus and SetTargetHealthStatus. This also reduces
  copying of the entire TargetHealth struct. It's also more convenient,
  as most of the time we just care about getting/setting health status
  field.