Fix (Cache) GetRole not found error message (#55444)
The web UI relies on error messages returned when roles do not
exist to be in a specific format to know when a user should be
reauthenticated. When switching the cache to use the new in memory
machinery the error message was not kept in sync with the
services/local implementation and resulted in weird errors in the
UI instead of detecting that a logout is required.

A test was added to ensure that the Cache and the Backend implementations
of GetRole return the same message when roles are not found.