Remove direct dependencies on lib/integrations/awsoidc
The awsoidc package is large, and pulls in several bloated aws
packages. This change decouples lib/srv and lib/reversetunnel from
the awsoidc package to remove indirect dependencies on awsoidc.

The EICE API was slightly tweaked to reduce the surface area at the
expense of basic tests. This code is largely deprecated due to the
strict limits enforced by AWS and likely headed for removal in a
future release.