make: turn *LDFLAGS properly evaluated
Currently, `make build/tsh` sets the `KUBECTL_SETVERSION` which is used
to set the kubectl version at compile time. This is used by `tsh kubectl
version` to print the kubectl version it ships.

The issue is `?=` isn't a lazy operator so it means all the flags are
evaluated once during the initial parsing of the makefile. It means that
when the phony executes, although `KUBECTL_SETVERSION` is set, during
the makefile parsing phase it wasn't so we don't pass the build flag.

This PR fixes that by moving the kubectl set code into the init parsing.
This also means we are setting the variable for all binaries but that
shouldn't have any meaningfull impact.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>