tools: Query teleport-private in addition for latest releases
Extend `query-latest` to query releases in `teleport-private` in
addition to the current `teleport` repository so that we can rebuild OCI
images released from `teleport-private`.

If the latest release for a version is in the `teleport-private`
repository, the version number printed will be prefixed with `private-`.

As `teleport-private` is a private repository, the environment variable
`GITHUB_TOKEN` must be set to a token with permissions to read the
releases of the `teleport-private` repository.

When testing locally, this can be done with:

    # may require `gh auth refresh` first
    export GITHUB_TOKEN=$(gh auth token)

assuming you have `gh` set up properly with credentials.