Handle no matching result for private separately.
Stop bundling all errors when looking up releases in teleport-private
and ignoring them all. Instead only treat no release matching the search
string as not an error as it is viable that there have not been any
private releases on a branch or minor version. Failure to query github
for private releases should be treated as an error, as not doing so may
mean we ignore a private release that actually is the latest.