tools: Query teleport-private in addition for latest releases (#55657)
* tools: Query teleport-private in addition for latest releases

Extend `query-latest` to query releases in `teleport-private` in
addition to the current `teleport` repository so that we can rebuild OCI
images released from `teleport-private`.

If the latest release for a version is in the `teleport-private`
repository, the version number printed will be prefixed with `private-`.

As `teleport-private` is a private repository, the environment variable
`GITHUB_TOKEN` must be set to a token with permissions to read the
releases of the `teleport-private` repository.

When testing locally, this can be done with:

    # may require `gh auth refresh` first
    export GITHUB_TOKEN=$(gh auth token)

assuming you have `gh` set up properly with credentials.

* Add repo name to "couldn't list releases" error message

* Handle no matching result for private separately.

Stop bundling all errors when looking up releases in teleport-private
and ignoring them all. Instead only treat no release matching the search
string as not an error as it is viable that there have not been any
private releases on a branch or minor version. Failure to query github
for private releases should be treated as an error, as not doing so may
mean we ignore a private release that actually is the latest.