[v16] kube: fix websocket fallback logic (#55733)
* kube: fix websocket fallback logic

Kubernetes 1.29 introduced support for the new exec over Websocket protocol.
In Kubernetes 1.29, this was an alpha feature which wasn't enabled by
default, but in Kubernetes 1.30 this feature flag was promoted to beta
stage and enabled by default. As such, we implemented the logic to
always prefer the Kubernetes exec over websocket if the Kubernetes
cluster supported it. The check that verified if we should prefer the
websocket was a simple Kubernetes version check and the condition was
`version >= 1.30.0`.

For OpenShift this check is not valid as they ship Kubernetes 1.30+ with
that feature disabled. Likely disabled at loadbalancer level. Given that
the version check wasn't reliable for all Kubernetes supported clusters,
this commit changes the logic to always use a fallback executor.

This executor initially tries to establish the websocket connection and
if it fails, it will fallback to the second executor. In our case, the
preferred executor is websocket and the secondary executor is SPDY.

Fixes #55695

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* fix data race from previous backport

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>