Refactor auth.Middleware to reduce dependencies on lib/auth (#55577)
* Refactor auth.Middleware to reduce dependencies on lib/auth

The base functionality of auth.Middleware has been moved to
authz.Middleware to allow decoupling lib/auth from more packages.
Most of the contents of auth.Middleware were better suited for
authz and did not depend on any auth specifics. The auth.Middleware
was not removed entirely, it still has some usefulness to the
auth servers. It now relies on the authz.Middleware to perform
the base functionality.

* fix: add docs to Middleware