Add an MWI section to the documentation (#55683)
Backports #55615

* Add a top-level MWI section to the docs

Move how-to guides related to Machine & Workload Identity into a
dedicated top-level documentation sidebar section so readers are aware
of this product as a top-level offering. This is part of a broader
initiative to add a top-level documentation section for each Teleport
product.

* Rename the mwi docs directory

Change the name to machine-workload-identity.

Responds to strideynet feedback