Attempt to make TestLockInForce less flaky (#55819) (#55845)
Updates the test to validate that the Lock created by the test is
observed by the LockWatcher before waiting for the session to be
terminated by the Lock. This should hopefully reduce flakes that
occur due to Lock propagation taking longer than the previous
timeout and seemed more robust than just bumping the existing
timeout by some additional margin.

Closes https://github.com/gravitational/teleport/issues/13770.