[AWSIC] Refactor IC Client and apply filters earlier (#6695)
This change pulls some complex operations out of the IC Client interface,
and converting their live implementations to helper functions defined in
terms of simpler operations from within the IC Client interface.

This both simplifies the client interface and reduces the amount of things
to mock for testing. I have moved the resulting the helper functions to the main `identitycenter` package.

This patch also pushes the application of account and group filters left,
back towards the initial import data fetch. This is in order to reduce the
amount of account and group loading the the AWS client needs to do when
only importing a subset of the IC accounts and groups.

Source-commit: gravitational/teleport.e@4827e20799cb6d5f64b93301d3fb70e5e87cb071
Source-pull-requests: gravitational/teleport.e#6695