make: Set kubectl version correctly for tsh (#6715)
Ensure `KUBECTL_SETVERSION` is propagated to the Go flags for building
tsh for the different ways that we can build tsh. We do this by removing
it from being a target-specific variable and setting it at the top-level
of the Makefile. Because this is always evaluated, rather than getting
the kubectl version by building and running a go program, use `go list`
instead.

Because the `Makefile` has a bare `export` keyword and `?=` is used to
set a bunch of variable, these variables get set the first time `make`
is invoked. In all the cases where `make` calls itself (such as in the
`all` target where `make` runs `$(MAKE) $(BINARIES)`), these variables
set with `?=` are inherited. If `KUBECTL_SETVERSION` is not set the
first time, its value will not be present in other flags. As
target-specific variables are only set for a that specific target, it is
likely the variable does not get set the first time make is invoked, and
setting it later when the target is actually called ends up doing
nothing.

Source-commit: gravitational/teleport.e@e68b6737cf9a6b2b22620dd0f72c42505be8431c
Source-pull-requests: gravitational/teleport.e#6715