ci: Shorten and regularlize AWS role-session-name (#6738)
Shorten one particular `role-session-name` in the `oci-image-build`
workflow as it had gotten too long (> 64 chars) causing the role
assumption step to fail. The proximate cause was using the git tag in
the role name which is quite long for the overnight builds on v18. The
ultimate cause was the run number reaching 1000, making the length 65
chars.

While here, regularize the role name in oci-image-{build,publish} in a
form that we should probably use consistently across all workflows. It
turns out the `github.run_number` is not very useful in workflows that
are called from other workflows as this is the run number of the
top-level workflow, not the actual workflow the assume-role step is in.

Using `github.run_id`-`github.job`-`github.run_attempt` suffix gives a
unique and traceable identifier such that if any suspicious actions in
AWS are noticed, they can be traced back to an actual run of a workflow
step. The suffix is 24 chars leaving plenty of room for a unique prefix
(typically gha-<workflow-name>-<purpose>).

The error in the GHA logs was:

    Error: Could not assume role with OIDC: 1 validation error detected: Value 'gha-oci-image-build-1010-18.0.0-alpha.3-nightly20250623-1-eebe9eb' at 'roleSessionName' failed to satisfy constraint: Member must have length less than or equal to 64

The other workflows will be reworked to use the same role-session-name
format when we do a round of hardening shortly.

Source-commit: gravitational/teleport.e@b7533173da3bea5abfa8cf918ece31603fd674e0
Source-pull-requests: gravitational/teleport.e#6738