Add `product` and `type` tags to docs pages
We are rolling out standard `product` and `type` frontmatter fields for
each docs page, letting us use this metadata to perform operations on
pages that fall under a certain type or product. For example, we could
generate product-specific landing pages using the `product` field.

This change acts in tandem with gravitational/docs-website#249, which
lints the frontmatter of each docs page to enforce the `product` and
`type` fields.

Note that some guides have "Reference" in the title but are really more
like conceptual guides, providing information about how aspects of
Teleport work instead of a comprehensive reference of possible inputs.

Note that this change marks category index pages as having type `other`.
It might make sense to add a type for `index`, but this is something we
can modify when we need to perform operations on category index pages.