Fix client_idle_timeout enforcement for desktop sessions (#56052)
We had been updating the client activity tracker any time a message
was sent from the client (browser or Teleport Connect) to the remote
host. This approach was fine for the original RDP implmentation, as
all messages sent in this direction were for user input (keypresses,
mouse movement, scroll wheel, etc), but it is insufficient with the
current RemoteFX implementation as there are some messages sent by
the remote Windows hosts which require client acknowledgements.
These acknowledgements were mistakenly being counted as client activity.

Closes #55691