Refactor cache to support storing resources in memory (#54799)
While the current cache storage is also in memory, it leverages
the memory backend which requires converting resources to and from
json. Marshaling json is suboptimal and is often a source of CPU
and memory consumption. The biggest problem with json marshaling
though is that it requires calling CheckAndSetDefaults. When
validations enforced in CASD become stricter it can leave caches
unable to become healthy when there are mixed versions within a
cluster as they all might have a slightly different view of what
should be allowed.

As a means to solve both problems, this changes the cache to store
resources received from the upstream directly in memory without
doing any conversion to json. There are two gotchas with this
approach, caching is no longer "free" and cloning of resources
must be done diligently to avoid races. Historically the cache
relied on the storage implementation in services/local to manage
persisting resources in the cache's backend.Memory. However, that
will no longer be the case as the cache needs to support storage
directly in memory. In order to reduce the burden this may pose on
developers the new collection machinery is simpler, and helpers
will be added on top of sortcache.SortCache to make a second
storage implementation trivial.

The changes here mark the existing collections machinery as legacy
and introduce new machinery to operate entirely in memory. The only
resources converted over are AccessRequests and HeadlessAuthentication.