Update database service stop db funcs
stopProxyingAndDeleteDatabase was called in only one place: unregisterDatabase.
It's more clear to just inline it rather than use a func named as
a synonym for "unregister".

Renamed stopUpstreamInventoryHeartbeat to
sendUpstreamInventoryStopHeartbeat. This is to disambiguate it from
stopHeartbeat, since it's actually sending an inventory control stream
message, not stopping a heartbeat directly.

Reworked stopHeartbeat to optionally delete the heartbeat as well,
either via sendUpstreamInventoryStopHeartbeat or directly from the
backend as a fallback. This way we don't pointlessly stop a heartbeat
that wasn't registered if an error occurred prior to registering the
heartbeat.