auditlog: introduce `USER_KIND_SYSTEM` for system roles
Teleport allows that system roles perform certain actions like creating,
updating or deleting users, roles and access lists when performed by the
Okta integration or creating, updating and deleting apps, kube, dbs
when performed by the discovery service.

When https://github.com/gravitational/teleport/pull/35881 was
implemented, it only contemplated two states: bot or human. If the
identity wasn't a bot, it was automatically tagged as a human.

This behavior is fine if we only emited audit logs for actions
performed by bots or users, but that's not the case. We also emit for
certain system actions.

When reading the audit log, one can see that the audit log is marked as
user although the username has the format: `<uuid>.<teleportClusterName>`.

This PR attempts to introduce a third user kind - system - that should
identity when the action was performed by a system component. This is a
requirement so that Identity Security can distinguish user actions and
system actions.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>