Fix error handling in firestore record creation (#56176)
The error checks were performed against error returned from
functions without assigning them to a local variable which resulted
in a return of nil, nil instead of returning the appropriate error.
In practice this isn't very likely to be hit since most keys should
be neither the legacy or broken format.