Generate the tsh CLI reference
Closes #47358

Run the docs generator introduced in #54394 for the `tsh` CLI reference.

Also remove the environment variable override for
`TELEPORT_LOGIN_BROWSER`, which is a hidden flag.

Note that the following hidden global environment variables are present
in the manually maintained guide but, because they are hidden, absent in
the generated guide:

- TELEPORT_LOGIN_BROWSER
- TELEPORT_USE_LOCAL_SSH_AGENT

The guide is also missing the `tsh puttyconfig` command, which is only
present when we build `tsh` for Windows. However, since VNET for SSH
fulfills much of the use case for `tsh puttyconfig`, and generating the
docs adds entries for around 33 more `tsh` commands, this is an
acceptable tradeoff.

Edit some help text to conform to the standards of the documentation.