ci: Add specificity to deploy-packages workflow run name (#6739)
Extend the `run-name` of the `deploy-packages` workflow to make it
easier to differentiate the different runs in the workflow list at
https://github.com/gravitational/teleport.e/actions/workflows/deploy-packages.yaml
This adds the build tag, environment and channels to the run name.

Currently it is hard to tell them all apart as only the package type
(apt or yum) is included in the run name.

Source-commit: gravitational/teleport.e@44869da9b40562d7c5e1c9bd01628489f0632e5a
Source-pull-requests: gravitational/teleport.e#6739