chore: remove `host_uuid` local storage compatibility layer (#56210)
This PR removes the compatibility layer introduced in Teleport 11 to support converting existing deployments that used storage to a the Kubernetes Secret storage used after Teleport 11.

This logic reads the data from the local file and re-stores it in Kubernetes secret if the secret doesn't have the `host_uuid` entry set. This behaviour should be sane because we re-get the secret to ensure it's not stored that but in reality this is no longer needed since Teleport 11 and should. be removed to ensure this isn't the cause of racy behaviour.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>