[AWSIC] Adds papercut protection to AWS update plugin handler (#6753)
Switching AWS IC Role Sync Modes from ALL (or empty, which is implicity
treated as ALL) to NONE leaves Teleport with orphan roles that are no
longer under IC control. It's not clear yet whether the "right thing"
to do is keep the roles, automatically delete them (which can be annoying
if the role is in use).

To prevent this happening until we have a good idea of the correct behaviour,
this pactch disallows changing from ALL to NONE.

See: Issue #6660

Source-commit: gravitational/teleport.e@83bcea3fd8d3e61f66e84ab695e1b2cea5535731
Source-pull-requests: gravitational/teleport.e#6753