Prevent missing locks in the connection monitor (#56414)
The order of operations by the monitor was susceptible to missing
locks because it only created the subscription for locks
_after_ checking if the known keys were in force. Any locks that
were created during the time at which the lock targets were in
force would never be seen. To avoid this the subscriptions are
always created first, then the lock in force check is performed
before proceeding.

Closes https://github.com/gravitational/teleport/issues/13770