[v18] Add an EKS disaster recovery guide (#55741)
* Add an EKS disaster recovery guide

When providing disaster recovery guidance for self-hosted Teleport
users, it is reasonable to begin with Amazon EKS users, since
self-hosted deployments on this platform are well supported. Provide a
guide for disaster recovery on EKS in the case of a regional outage.

This change adds the guide to the `helm-deployments` section, which
contains other guidance for self-hosted Teleport clusters on Kubernetes.
It changes the section index page to use `DocCardList` instead of an
incomplete, manually maintained list of guides.

While initial plans for this guide included a `tctl get all` discussion,
we don't recommend this approach because it doesn't back up all backend
data, so this change does not include this discussion in guide. Instead,
add a link to the Backup and Restore guide, which does discuss this
command.

Related changes:

- Extract DNS information in the `kubernetes-cluster.mdx` guide into a
  partial
- Fix an outdated DynamoDB docs link in the Backup and Restore guide

* Respond to pschisa feedback

- Mention the TTL of the Proxy Service DNS records.
- Make the change freeze discussion an optional section under "Guidance"
  instead of part of the main procedure, since shutting down a cluster
  makes changes impossible.
- Note how a full regional outage would affect the user's ability to
  execute the "Stop the existing cluster" step.
- Use `helm uninstall` instead of deleting deployments in order to stop
  the existing cluster.
- Add a note about recovery point objective.

* Respond to hugoShaka feedback

Describe restarting the Proxy Service in relation to gRPC connections as
a way to prevent agents from remaining connected to affected Proxy
Service instances.