ci: Update GHA cron jobs for v18 release (#6796)
Update the GitHub Actions cron jobs now that v18 has been released. This
means rebuilding and monitoring the OCI images for that branch and
building the buildbox for it nightly.

Normally we would remove v15 from all of these, but I've left it in for
the OCI monitoring and rebuild as we still have people on v15 on cloud.
Once automatic updates are fully deployed and people updated to
supported versions, we can remove v15 from the OCI cron jobs.

I have removed it from the buildbox rebuild as that's pretty stable and
rarely gets rebuilt due to centos7 being EOL.

Source-commit: gravitational/teleport.e@84a1df545a3bd2b40b771a9d48e1083b71c54b37
Source-pull-requests: gravitational/teleport.e#6796