[v16] Add a guide to Proxy Service traffic separation (#56436)
* Add a guide to Proxy Service traffic separation

Closes #27885

Add a guide to the Self-Hosting Teleport section laying out the options
described in #27885 for separating internal and external Proxy Service
traffic.

* Respond to pschisa feedback

- Make it more explicit that the guide applies to self-hosted clusters.
- Mention Managed Updates v2 instead of package managers.
- Fix example ports in the TELEPORT_TUNNEL_PUBLIC_ADDR section.
- Mention DNS in the "Multiple public addresses" section

* Respond to feedback

Edit the Proxy Service traffic separation guide to include a strong
recommendation to use the split DNS approach. Edit explanations of other
approaches to frame them as contingencies for if a cluster doesn't meet
the requirements of the split DNS approach.