Remove lib/pam dependency from lib/service/servicecfg
The servicecfg package is meant to be a relatively lightweight
package, with as few dependencies as possible, so that it may
be consumed by teleport and client tools alike. The pam validation
that was being performed when parsing the pam configuration results
in servicecfg being transitively dependant on CGO, which is not
lightweight.

This moves the pam validation into the teleport process start up
code along side similar validations. As a result of this change
tctl no longer has a dependency on lib/pam.