Stop using keystore.NewSoftwareKeystoreForTests (#6844)
The test helper provides no more functionality than calling
keystore.NewManager does at the expense of causing testing symbols
to appear in production binaries. In an effort to remove the helper
it's use in enterprise is first removed so that it doesn't break
builds when the function is deleted entirely.

Source-commit: gravitational/teleport.e@a7e790a79d751af6f3b31334e358d70ea4e6b577
Source-pull-requests: gravitational/teleport.e#6844