MWI: Fix bound keypair initial join secret field name (#56603)
* MWI: Fix bound keypair initial join secret field name

The `initial_join_secret` field was not given a proper YAML field
name and was rendering as `initialjoinsecret`. Additionally, we've
tried to standardize on referring to this field as "the registration
secret", so this renames the field to match new terminology.

This hopefully does not count as a breaking change as registration
secret functionality has not been made available in a release.

* Rename to `registration_secret`