oprt: Fix locking with Kubernetes lease (#6760)
Return an error when trying to get the serialization lock if the context
is canceled, instead of falling through and allowing work to proceed
without the lock.

Fix `GetSafeName` to return the actual safe name it generates rather
than its parameter unchanged.

Add a log when we start the leader election, including our identity to
make it easier to correlate the logs with the k8s lease.

Source-commit: gravitational/teleport.e@b59b6a1ea887fd4c0ced3fa1a1b13373d31cd67c
Source-pull-requests: gravitational/teleport.e#6760