Align Identity Center APIs with other resources
This is the first in a series of changes to migrate the identity
center APIs away from `lib/utils/pagination` with the end goal
of removing the pagination package entirely.

## Why?

Pagination bugs have been a major problem historically. Having
multiple implementations for iterating pages of resources increases
the chances of a bug. By consolidating our pagination APIs to a
single shape, we can reduce the number of hand rolled implementations
to consume an entire set of pages.

https://github.com/gravitational/teleport/pull/56418 and
https://github.com/gravitational/teleport/pull/56581 have recently
consolidated _most_ consumption of pagination APIs to a single
helper in `api/utils/clientutils`. However, due to the use of
the pagination.NextKey by the Identity Center APIs they need either
special cases, or their own special implementation.

The IdentityCenterAccount and IdentityCenterAccountAssignment wrapper
types are exposed outside of where they are required. They were
added solely for the unified resource cache, and should not have
been included in the APIs for the resources. The unified resource
cache no longer uses the wrappers but they cannot be removed until
the APIs no longer reference them.

## How?

These APIs are consumed in both OSS and Enterprise code which means
they cannot be changed directly without breaking builds. The plan
is to provide equivalent methods with different signatures to allow
the transition to occur over a series of changes.